Add support for configuration
This configuration allow the administrator to specify if decryption is authorized or not and to provide the encryption key from the environment. Decryption is disabled when encryption is provided to avoid security issues (ex: production secret decrypted).
It is also possible to select the default algorithm and mode used to encrypt/decrypt the properties.
Finally, the UI will adapt based on a new endpoint returning the configuration.